chore: refactor Substrait consumer's "rename_field" and implement the rest of types (#16345)
* refactor Substrait consumer's "rename_field" and implement the rest of types

The rename_field is a bit confusing with it's "rename_self" parameter. Also there are times when one wants to just rename a data type.
And I'd like to reuse the same code within our codebase, since with Substrait the case of renaming a field/type comes up a bunch, so I'd like to make these functions "pub".

And lastly, this adds support for all list types, as well as dict/ree/union types. Dunno how necessary those are, but seems right to support them still.

* datatype => data_type

* fix clippy