Switch from xz2 to liblzma to reduce duplicate dependencies (#17509)
## Which issue does this PR close?

Closes https://github.com/apache/datafusion/issues/15342

## Rationale for this change

Reduces the duplicate dependencies.

We currently depend on bzip2 in two different ways. In attempting to
reduce this, I needed to update `async-compression` which caused two
different libraries to link to the system lzma library. This is not
allowed in rust.

This PR updates avro-rs, but we cannot merge this PR until that crate
merges https://github.com/apache/avro-rs/pull/284 and we remove the
crates.io patch this PR contains.

## What changes are included in this PR?

Update avro-rs and switch from the unmaintained xz2 crate to liblzma.

## Are these changes tested?

Unit tests. 

## Are there any user-facing changes?

None. This is simply a dependency update to a more recent crate.