Refactor InListExpr to support structs by re-using existing hashing infrastructure  (#18449)
This PR is part of an EPIC to push down hash table references from
HashJoinExec into scans. The EPIC is tracked in
https://github.com/apache/datafusion/issues/17171.

A "target state" is tracked in
https://github.com/apache/datafusion/pull/18393.
There is a series of PRs to get us to this target state in smaller more
reviewable changes that are still valuable on their own:
- https://github.com/apache/datafusion/pull/18448
- (This PR): https://github.com/apache/datafusion/pull/18449 (depends on
https://github.com/apache/datafusion/pull/18448)
- https://github.com/apache/datafusion/pull/18451

- Enhance InListExpr to efficiently store homogeneous lists as arrays
and avoid a conversion to Vec<PhysicalExpr>
  by adding an internal InListStorage enum with Array and Exprs variants
- Re-use existing hashing and comparison utilities to support Struct
arrays and other complex types
- Add public function `in_list_from_array(expr, list_array, negated)`
for creating InList from arrays

Although the diff looks large most of it is actually tests and docs. I
think the actual code change is a negative LOC change, or at least
negative complexity (eliminates a trait, a macro, matching on data
types).

---------

Co-authored-by: David Hewitt <mail@davidhewitt.dev>
Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>
(cherry picked from commit 486c5d860b02fba8aca53a10fca918d3e23e3d26)