Push down InList or hash table references from HashJoinExec depending on the size of the build side (#18393)
This PR is part of an EPIC to push down hash table references from
HashJoinExec into scans. The EPIC is tracked in
https://github.com/apache/datafusion/issues/17171.

A "target state" is tracked in
https://github.com/apache/datafusion/pull/18393 (*this PR*).
There is a series of PRs to get us to this target state in smaller more
reviewable changes that are still valuable on their own:
- https://github.com/apache/datafusion/pull/18448
- https://github.com/apache/datafusion/pull/18449 (depends on
https://github.com/apache/datafusion/pull/18448)
- https://github.com/apache/datafusion/pull/18451

As those are merged I will rebase this PR to keep track of the
"remaining work", and we can use this PR to explore big picture ideas or
benchmarks of the final state.

(cherry picked from commit c0e8bb501a7b62bb40c87edf16d46ee53bdaff73)